Update angular-table package and dependencies versions
The version for angular-table package has been updated from 17.0.2 to 17.1.1. Similarly, the versions for all dependencies in the package.json file including, but not limited to, "@angular/animations", "@angular/common", and "@angular/core" have also been updated to align with the new version. In addition, significant updates were made to the CSS styles for table components.